+++ main.go
@@ -9,9 +9,7 @@ import (
 	"syscall"
 
 	"git.sr.ht/~jrswab/akordo/controller"
-	"git.sr.ht/~jrswab/akordo/plugins"
-	"git.sr.ht/~jrswab/akordo/roles"
-	"git.sr.ht/~jrswab/akordo/xp"
+	"git.sr.ht/~jrswab/akordo/load"
 	"github.com/bwmarrin/discordgo"
 )
 
@@ -20,33 +18,6 @@ var token string
 func init() {
 	flag.StringVar(&token, "t", "", "Bot Token")
 	flag.Parse()
-}
-
-func loadSavedData(sd *controller.SessionData) {
-	// Load saved XP data into the struct created by NewSessionData
-	if _, err := os.Stat(xp.XpFile); err == nil {
-		if err := sd.XP.LoadXP(xp.XpFile); err != nil {
-			log.Fatalf("error loading xp data file: %s", err)
-		}
-	}
-	// Load saved role data into the struct created by NewSessionData
-	if _, err := os.Stat(xp.AutoRankFile); err == nil {
-		if err := sd.XP.LoadAutoRanks(xp.AutoRankFile); err != nil {
-			log.Fatalf("error loading role file: %s", err)
-		}
-	}
-	// Load saved self assign role data into the struct created by NewSessionData
-	if _, err := os.Stat(roles.SelfAssignFile); err == nil {
-		if err := sd.Roles.LoadSelfAssignRoles(roles.SelfAssignFile); err != nil {
-			log.Fatalf("error loading self assign role file: %s", err)
-		}
-	}
-	// Load saved banned word data into the struct created by NewSessionData
-	if _, err := os.Stat(plugins.BannedWordsPath); err == nil {
-		if err := sd.Blacklist.LoadBannedWordList(plugins.BannedWordsPath); err != nil {
-			log.Fatalf("error loading banned words file: %s", err)
-		}
-	}
 }
 
 func main() {
@@ -67,7 +38,7 @@ func main() {
 	// Create a the custom controller to pass data to ReceiveMessage and the plugins
 	sd := controller.NewSessionData(sess)
 
-	loadSavedData(sd)
+	load.SavedData(sd)
 
 	// start the Goroutine to automatically save earned XP
 	go sd.XP.AutoSaveXP()