+++ plugins/bannedWords.go
@@ -12,7 +12,7 @@ import (
 )
 
 // BannedWordsPath is the path to the banned words json file
-const BannedWordsPath string = "bannedWords.json"
+const BannedWordsPath string = "data/bannedWords.json"
 
 // Blacklist contains the data needed to execute the method functionality
 type Blacklist struct {